create pihole.yaml
This is not using kubernetes secrets! 
Its a stateless approach but pihole removed the ability to provide an adlists.list so this part is currently not working, i already found some scripts to alter the database but maybe a statefull approach would be better than that. As i havent fully compared both possible solutions the current approach is to just use the list provided by default.